api_gateway/requirements.txt
# FastAPI 및 ASGI 서버
fastapi==0.111.0
uvicorn[standard]==0.29.0

# 파일 업로드 처리
python-multipart==0.0.9

# Kafka 클라이언트
kafka-python==2.0.2

# MinIO 클라이언트
minio==7.2.7

# PostgreSQL 연동 (ORM + 드라이버)
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# 인증 및 보안 (JWT, 비밀번호 해싱)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 데이터 유효성 검사 (Pydantic은 FastAPI의 의존성 라이브러리지만, 명시적으로 버전 관리)
pydantic[email]==2.7.1

# 환경 변수 관리
python-dotenv==1.0.1

# WebSocket 지원
websockets==12.0

# OAuth2 (FastAPI의 의존성 라이브러리지만, 명시적으로 버전 관리)
python-jose==3.3.0

# Test dependencies
pytest==8.2.2
httpx==0.27.0
pytest-mock==3.12.0